+++ com.sparta.logistics.delivery-service/src/main/java/com/sparta/logistics/delivery_service/infrastructure/client/HubRouteClient.java
@@ -1,6 +1,6 @@
 package com.sparta.logistics.delivery_service.infrastructure.client;
 
-import com.sparta.logistics.delivery_service.application.dto.response.HubRouteListResponseDto;
+import com.sparta.logistics.delivery_service.application.dto.HubRouteListResponseDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
